backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
playwright==1.42.0
beautifulsoup4==4.12.3
python-multipart==0.0.9
google-genai==0.3.0
pydantic==2.6.3
pypdf==4.1.0
aiofiles==23.2.1
pdf2image==1.17.0
pytesseract==0.3.10
python-dotenv==1.0.1
aiohttp==3.9.3
langchain-community==0.0.28
langsmith==0.0.87
python-docx==1.1.0
requests==2.31.0
openpyxl==3.1.2
striprtf==0.0.26
xlrd==2.0.1
